feat: Remove old collection UI style and feature gating
