+++ packages/frontend/@n8n/i18n/src/locales/en.json
@@ -997,11 +997,9 @@
 	"codeNodeEditor.parameters": "Parameters",
 	"codeNodeEditor.optional": "optional",
 	"codeNodeEditor.defaultsTo": "Defaults to {default}.",
-	"collectionParameter.choose": "Choose...",
 	"collectionParameter.addItem": "Add item",
 	"collectionParameter.deleteItem": "Delete",
 	"collectionParameter.dragItem": "Drag to reorder",
-	"collectionParameter.noProperties": "No properties",
 	"collectionParameter.allOptionsAdded": "All options have been added",
 	"communityNodeFooter.legacy": "Legacy",
 	"communityNodeFooter.manage": "Manage",
@@ -1679,7 +1677,6 @@
 	"fixedCollectionParameter.addParameter": "Add {parameter}",
 	"fixedCollectionParameter.allOptionsAdded": "All options have been added",
 	"fixedCollectionParameter.addField": "Add Field",
-	"fixedCollectionParameter.choose": "Choose...",
 	"fixedCollectionParameter.currentlyNoItemsExist": "Currently no items exist",
 	"fixedCollectionParameter.deleteItem": "Delete item",
 	"fixedCollectionParameter.deleteParameter": "Delete {parameter}",
