--- packages/frontend/editor-ui/src/features/ndv/parameters/components/Collection/CollectionParameterLegacy.test.ts
@@ -1,619 +1,0 @@
-import { createComponentRenderer } from '@/__tests__/render';
-import { SETTINGS_STORE_DEFAULT_STATE } from '@/__tests__/utils';
-import CollectionParameterLegacy, { type Props } from './CollectionParameterLegacy.vue';
-import { STORES } from '@n8n/stores';
-import { createTestingPinia } from '@pinia/testing';
-import userEvent from '@testing-library/user-event';
-import { setActivePinia } from 'pinia';
-import { nextTick } from 'vue';
-import { flushPromises } from '@vue/test-utils';
-
-// Instantiates a store that derives the workflow id from the route. These tests run
-// without a router, so resolve the id directly.
-vi.mock('@/app/composables/useWorkflowId', async () => {
-	const { computed } = await import('vue');
-	return {
-		useWorkflowId: () => computed(() => ''),
-		useRouteWorkflowId: () => computed(() => ''),
-	};
-});
-
-// Controllable active node + gateway lookups so the AI Gateway hiding path can be
-// exercised. Defaults match the no-active-node behaviour the other tests rely on.
-let mockActiveNode: unknown = null;
-const mockIsNodePropertyHidden = vi.fn((_node: unknown, _param: string) => false);
-
-vi.mock('@/features/ndv/shared/ndv.store', async (importOriginal) => {
-	const actual = await importOriginal<typeof import('@/features/ndv/shared/ndv.store')>();
-	return {
-		...actual,
-		injectNDVStore: () => ({
-			value: {
-				get activeNode() {
-					return mockActiveNode;
-				},
-			},
-		}),
-	};
-});
-
-vi.mock('@/app/stores/aiGateway.store', () => ({
-	useAiGatewayStore: () => ({ isNodePropertyHidden: mockIsNodePropertyHidden }),
-}));
-
-describe('CollectionParameterLegacy.vue', () => {
-	const pinia = createTestingPinia({
-		initialState: {
-			[STORES.SETTINGS]: {
-				settings: SETTINGS_STORE_DEFAULT_STATE.settings,
-			},
-		},
-	});
-	setActivePinia(pinia);
-
-	const baseProps: Props = {
-		parameter: {
-			displayName: 'Additional Fields',
-			name: 'additionalFields',
-			placeholder: 'Add Field',
-			type: 'collection',
-			default: {},
-			options: [
-				{
-					displayName: 'Currency',
-					name: 'currency',
-					type: 'string',
-					default: 'USD',
-				},
-				{
-					displayName: 'Value',
-					name: 'value',
-					type: 'number',
-					default: 0,
-				},
-			],
-		},
-		path: 'parameters.additionalFields',
-		nodeValues: {
-			parameters: {
-				additionalFields: {},
-			},
-		},
-		values: {},
-		isReadOnly: false,
-		isNested: false,
-	};
-
-	const renderComponent = createComponentRenderer(CollectionParameterLegacy, {
-		props: baseProps,
-	});
-
-	describe('Rendering', () => {
-		it('renders the component', () => {
-			const { container } = renderComponent();
-			expect(container).toBeInTheDocument();
-		});
-
-		it('shows "no properties" message when empty and not nested', () => {
-			const { getByText } = renderComponent({
-				props: {
-					...baseProps,
-					values: {},
-				},
-			});
-
-			expect(getByText('No properties')).toBeInTheDocument();
-		});
-
-		it('does not show "no properties" message when nested', () => {
-			const { queryByText } = renderComponent({
-				props: {
-					...baseProps,
-					values: {},
-					isNested: true,
-				},
-			});
-
-			expect(queryByText('No properties')).not.toBeInTheDocument();
-		});
-	});
-
-	describe('Adding items - Single option', () => {
-		it('renders a button when only one option is available', () => {
-			const singleOptionProps: Props = {
-				...baseProps,
-				parameter: {
-					...baseProps.parameter,
-					options: [
-						{
-							displayName: 'Currency',
-							name: 'currency',
-							type: 'string',
-							default: 'USD',
-						},
-					],
-				},
-			};
-
-			const { getByTestId } = renderComponent({
-				props: singleOptionProps,
-			});
-
-			expect(getByTestId('collection-parameter-add')).toBeInTheDocument();
-		});
-
-		it('emits valueChanged event when clicking single option button', async () => {
-			const singleOptionProps: Props = {
-				...baseProps,
-				parameter: {
-					...baseProps.parameter,
-					options: [
-						{
-							displayName: 'Currency',
-							name: 'currency',
-							type: 'string',
-							default: 'USD',
-						},
-					],
-				},
-			};
-
-			const { getByTestId, emitted } = renderComponent({
-				props: singleOptionProps,
-			});
-
-			const addButton = getByTestId('collection-parameter-add');
-			await userEvent.click(addButton);
-
-			expect(emitted('valueChanged')).toBeDefined();
-			expect(emitted('valueChanged')?.[0]).toEqual([
-				{
-					name: 'parameters.additionalFields.currency',
-					value: 'USD',
-				},
-			]);
-		});
-	});
-
-	describe('Adding items - Multiple options', () => {
-		it('renders a select dropdown when multiple options are available', () => {
-			const { getByRole } = renderComponent();
-
-			const select = getByRole('combobox');
-			expect(select).toBeInTheDocument();
-		});
-
-		it('shows all available options in dropdown', () => {
-			const { getAllByTestId } = renderComponent();
-
-			const options = getAllByTestId('collection-parameter-option');
-			expect(options.length).toBe(2);
-			expect(options[0]).toHaveTextContent('Currency');
-			expect(options[1]).toHaveTextContent('Value');
-		});
-
-		it('emits valueChanged event when selecting an option', async () => {
-			const { getByRole } = renderComponent();
-
-			const select = getByRole('combobox');
-			await userEvent.click(select);
-
-			// Simulate selecting the first option
-			const firstOption = select.querySelector('option[value="currency"]') as HTMLOptionElement;
-			if (firstOption) {
-				await userEvent.selectOptions(select, firstOption);
-			}
-
-			await nextTick();
-
-			// The component should emit the event (exact behavior depends on N8nSelect implementation)
-			// This test verifies the structure is correct
-			expect(select).toBeInTheDocument();
-		});
-
-		it('filters out already added options from dropdown', async () => {
-			const { getAllByTestId } = renderComponent({
-				props: {
-					...baseProps,
-					values: {
-						currency: 'USD',
-					},
-					nodeValues: {
-						parameters: {
-							additionalFields: {
-								currency: 'USD',
-							},
-						},
-					},
-				},
-			});
-			await flushPromises();
-
-			// With 2 total options and 1 added, the select dropdown should show only the remaining option
-			const options = getAllByTestId('collection-parameter-option');
-			expect(options).toHaveLength(1);
-			expect(options[0]).toHaveTextContent('Value');
-		});
-	});
-
-	describe('Read-only mode', () => {
-		it('does not render add controls when isReadOnly is true', async () => {
-			const { queryByTestId, queryByRole } = renderComponent({
-				props: {
-					...baseProps,
-					isReadOnly: true,
-				},
-			});
-			await flushPromises();
-
-			expect(queryByTestId('collection-parameter-add')).not.toBeInTheDocument();
-			expect(queryByRole('combobox')).not.toBeInTheDocument();
-		});
-	});
-
-	describe('Add controls visibility', () => {
-		it('hides add controls when all options are added (multiple options)', async () => {
-			const { queryByTestId, queryByRole } = renderComponent({
-				props: {
-					...baseProps,
-					values: {
-						currency: 'USD',
-						value: 100,
-					},
-					nodeValues: {
-						parameters: {
-							additionalFields: {
-								currency: 'USD',
-								value: 100,
-							},
-						},
-					},
-				},
-			});
-			await flushPromises();
-
-			expect(queryByTestId('collection-parameter-add')).not.toBeInTheDocument();
-			expect(queryByRole('combobox')).not.toBeInTheDocument();
-		});
-
-		it('hides add button when single option is already added', async () => {
-			const { queryByTestId } = renderComponent({
-				props: {
-					...baseProps,
-					parameter: {
-						...baseProps.parameter,
-						options: [
-							{
-								displayName: 'Currency',
-								name: 'currency',
-								type: 'string',
-								default: 'USD',
-							},
-						],
-					},
-					values: {
-						currency: 'USD',
-					},
-					nodeValues: {
-						parameters: {
-							additionalFields: {
-								currency: 'USD',
-							},
-						},
-					},
-				},
-			});
-			await flushPromises();
-
-			expect(queryByTestId('collection-parameter-add')).not.toBeInTheDocument();
-		});
-
-		it('hides add controls when parameter.options is empty', async () => {
-			const { queryByTestId, queryByRole } = renderComponent({
-				props: {
-					...baseProps,
-					parameter: {
-						...baseProps.parameter,
-						options: [],
-					},
-				},
-			});
-			await flushPromises();
-
-			expect(queryByTestId('collection-parameter-add')).not.toBeInTheDocument();
-			expect(queryByRole('combobox')).not.toBeInTheDocument();
-		});
-
-		it('hides add controls when parameter.options is undefined', async () => {
-			const { queryByTestId, queryByRole } = renderComponent({
-				props: {
-					...baseProps,
-					parameter: {
-						...baseProps.parameter,
-						options: undefined,
-					},
-				},
-			});
-			await flushPromises();
-
-			expect(queryByTestId('collection-parameter-add')).not.toBeInTheDocument();
-			expect(queryByRole('combobox')).not.toBeInTheDocument();
-		});
-
-		it('does not disable select dropdown when options remain available', () => {
-			const { getByRole } = renderComponent();
-
-			const select = getByRole('combobox');
-			expect(select).not.toBeDisabled();
-		});
-	});
-
-	describe('Collection values', () => {
-		it('handles INodePropertyCollection options', async () => {
-			const collectionProps: Props = {
-				...baseProps,
-				parameter: {
-					...baseProps.parameter,
-					options: [
-						{
-							name: 'collection1',
-							displayName: 'Collection 1',
-							values: [
-								{
-									displayName: 'Field 1',
-									name: 'field1',
-									type: 'string',
-									default: '',
-								},
-							],
-						},
-					],
-				},
-			};
-
-			const { container } = renderComponent({
-				props: collectionProps,
-			});
-			await flushPromises();
-
-			expect(container).toBeInTheDocument();
-		});
-
-		it('handles mixed INodeProperties and INodePropertyCollection options', async () => {
-			const mixedProps: Props = {
-				...baseProps,
-				parameter: {
-					...baseProps.parameter,
-					options: [
-						{
-							displayName: 'Simple Field',
-							name: 'simpleField',
-							type: 'string',
-							default: '',
-						},
-						{
-							name: 'collection1',
-							displayName: 'Collection 1',
-							values: [
-								{
-									displayName: 'Field 1',
-									name: 'field1',
-									type: 'string',
-									default: '',
-								},
-							],
-						},
-					],
-				},
-			};
-
-			const { container } = renderComponent({
-				props: mixedProps,
-			});
-			await flushPromises();
-
-			expect(container).toBeInTheDocument();
-		});
-	});
-
-	describe('Value handling', () => {
-		it('adds newly added parameter to tracking set', async () => {
-			const singleOptionProps: Props = {
-				...baseProps,
-				parameter: {
-					...baseProps.parameter,
-					options: [
-						{
-							displayName: 'Currency',
-							name: 'currency',
-							type: 'string',
-							default: 'USD',
-						},
-					],
-				},
-			};
-
-			const { getByTestId, emitted } = renderComponent({
-				props: singleOptionProps,
-			});
-			await flushPromises();
-
-			const addButton = getByTestId('collection-parameter-add');
-			await userEvent.click(addButton);
-			await flushPromises();
-
-			// Verify event was emitted
-			expect(emitted('valueChanged')).toBeDefined();
-		});
-
-		it('clears selection after adding an option', async () => {
-			const singleOptionProps: Props = {
-				...baseProps,
-				parameter: {
-					...baseProps.parameter,
-					options: [
-						{
-							displayName: 'Currency',
-							name: 'currency',
-							type: 'string',
-							default: 'USD',
-						},
-					],
-				},
-			};
-
-			const { getByTestId } = renderComponent({
-				props: singleOptionProps,
-			});
-			await flushPromises();
-
-			const addButton = getByTestId('collection-parameter-add');
-			await userEvent.click(addButton);
-
-			await nextTick();
-			await flushPromises();
-
-			// Button should still be present (selection cleared internally)
-			expect(addButton).toBeInTheDocument();
-		});
-	});
-
-	describe('Placeholder text', () => {
-		it('uses parameter placeholder when available', async () => {
-			const { getByTestId } = renderComponent({
-				props: {
-					...baseProps,
-					parameter: {
-						...baseProps.parameter,
-						options: [
-							{
-								displayName: 'Currency',
-								name: 'currency',
-								type: 'string',
-								default: 'USD',
-							},
-						],
-					},
-				},
-			});
-			await flushPromises();
-
-			const addButton = getByTestId('collection-parameter-add');
-			expect(addButton).toHaveTextContent('Add Field');
-		});
-	});
-
-	describe('Nested collections', () => {
-		it('renders correctly when isNested is true', async () => {
-			const { container } = renderComponent({
-				props: {
-					...baseProps,
-					isNested: true,
-					values: {
-						currency: 'USD',
-					},
-					nodeValues: {
-						parameters: {
-							additionalFields: {
-								currency: 'USD',
-							},
-						},
-					},
-				},
-			});
-			await flushPromises();
-
-			expect(container).toBeInTheDocument();
-		});
-	});
-
-	describe('AI Gateway hidden properties', () => {
-		afterEach(() => {
-			mockActiveNode = null;
-			mockIsNodePropertyHidden.mockReset();
-			mockIsNodePropertyHidden.mockReturnValue(false);
-		});
-
-		it('removes properties the store reports as hidden', async () => {
-			mockIsNodePropertyHidden.mockImplementation((_node, param) => param === 'value');
-
-			const { getAllByTestId } = renderComponent();
-			await flushPromises();
-
-			const options = getAllByTestId('collection-parameter-option');
-			expect(options).toHaveLength(1);
-			expect(options[0]).toHaveTextContent('Currency');
-		});
-
-		it('keeps properties the store does not hide', async () => {
-			mockIsNodePropertyHidden.mockReturnValue(false);
-
-			const { getAllByTestId } = renderComponent();
-			await flushPromises();
-
-			expect(getAllByTestId('collection-parameter-option')).toHaveLength(2);
-		});
-
-		it('removes hidden collection-type options', async () => {
-			mockIsNodePropertyHidden.mockImplementation((_node, param) => param === 'nestedCollection');
-
-			const { getAllByTestId } = renderComponent({
-				props: {
-					...baseProps,
-					parameter: {
-						...baseProps.parameter,
-						options: [
-							{
-								displayName: 'Currency',
-								name: 'currency',
-								type: 'string',
-								default: 'USD',
-							},
-							{
-								name: 'nestedCollection',
-								displayName: 'Nested Collection',
-								values: [
-									{
-										displayName: 'Field 1',
-										name: 'field1',
-										type: 'string',
-										default: '',
-									},
-								],
-							},
-						],
-					},
-				},
-			});
-			await flushPromises();
-
-			const options = getAllByTestId('collection-parameter-option');
-			expect(options).toHaveLength(1);
-			expect(options[0]).toHaveTextContent('Currency');
-		});
-	});
-
-	describe('hideDelete prop', () => {
-		it('passes hideDelete to ParameterInputList', async () => {
-			const { container } = renderComponent({
-				props: {
-					...baseProps,
-					hideDelete: true,
-					values: {
-						currency: 'USD',
-					},
-					nodeValues: {
-						parameters: {
-							additionalFields: {
-								currency: 'USD',
-							},
-						},
-					},
-				},
-			});
-			await flushPromises();
-
-			expect(container).toBeInTheDocument();
-		});
-	});
-});
